requirements.txt
requests
python-binance
python-dotenv
pandas
cryptography
aiohttp
